Update Qdrant to 1.1.4
