+++ pyproject.toml
@@ -27,7 +27,7 @@ pandas = {version = ">=1.1.0", optional = true }
 elasticsearch = {version = "7.10.1", optional = true }
 smart-open = {version = ">=6.3.0", extras = ["s3"], optional = true}
 jina-hubble-sdk = {version = ">=0.34.0", optional = true}
-qdrant-client = {version = ">=1.1.1", python = "<3.12", optional = true }
+qdrant-client = {version = ">=1.1.4", python = "<3.12", optional = true }
 
 [tool.poetry.extras]
 proto = ["protobuf", "lz4"]
