+++ tests/index/qdrant/test_persist_data.py
@@ -6,7 +6,7 @@ from docarray import BaseDoc
 from docarray.index import QdrantDocumentIndex
 from docarray.typing import NdArray
 
-import qdrant_client
+from .fixtures import qdrant_config, qdrant  # ignore: type[import]
 
 pytestmark = [pytest.mark.slow, pytest.mark.index]
 
@@ -18,18 +18,6 @@ class SimpleDoc(BaseDoc):
 class NestedDoc(BaseDoc):
     d: SimpleDoc
     tens: NdArray[50]  # type: ignore[valid-type]
-
-
-@pytest.fixture
-def qdrant_config():
-    return QdrantDocumentIndex.DBConfig()
-
-
-@pytest.fixture
-def qdrant():
-    """This fixture takes care of removing the collection before each test case"""
-    client = qdrant_client.QdrantClient('http://localhost:6333')
-    client.delete_collection(collection_name='documents')
 
 
 def test_persist_and_restore(qdrant_config, qdrant):
