+++ tests/index/qdrant/test_text_search.py
@@ -1,5 +1,3 @@
-import pytest
-import qdrant_client
 import numpy as np
 
 from pydantic import Field
@@ -8,22 +6,12 @@ from docarray import BaseDoc
 from docarray.index import QdrantDocumentIndex
 from docarray.typing import NdArray
 
+from .fixtures import qdrant_config, qdrant  # ignore: type[import]
+
 
 class SimpleDoc(BaseDoc):
     embedding: NdArray[10] = Field(dim=1000)  # type: ignore[valid-type]
     text: str
-
-
-@pytest.fixture
-def qdrant_config():
-    return QdrantDocumentIndex.DBConfig()
-
-
-@pytest.fixture
-def qdrant():
-    """This fixture takes care of removing the collection before each test case"""
-    client = qdrant_client.QdrantClient('http://localhost:6333')
-    client.delete_collection(collection_name='documents')
 
 
 def test_text_search(qdrant_config, qdrant):
